+++ src/hooks/useUser.ts
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { UserContext } from "../providers/_UserProvider";
+import { UserContext } from "../providers/UserProvider";
 
 export const useUser = () => useContext(UserContext);
